Continue sync and refactor
