+++ src/Actions/index.js
@@ -1,7 +1,14 @@
 // @flow
-
-import type { Dispatch, GetState, MessageType, Id } from '../FlowTypes/';
+import { WIDGET_ID } from '../Constants/index';
+import type { MessageType, Id, Dispatch, GetState } from '../FlowTypes/';
 import SLACK_API from '../Utils/Api';
+// TODO: Needs to be an environmental variable
+const PATH = 'http://localhost:4000/';
+
+/* eslint func-names: ["error", "never"] */
+function fetchRequest(path) {
+  return fetch(path).then(response => response.json());
+}
 
 export function connectWithSlack() {
   return {
@@ -9,6 +16,37 @@ export function connectWithSlack() {
   };
 }
 
+export function fetchChannels() {
+  return async function (dispatch: Dispatch) {
+    // const response = await fetchRequest(`${PATH}channels`);
+    const channels = await SLACK_API.fetchRequestChannels();
+    dispatch({
+      channels,
+      type: 'RECEIVED_CHANNEL_LIST',
+    });
+  };
+}
+
+
+
+
+export function fetchMessagesForChannel(channel: string) {
+  return async function (dispatch: Dispatch, getState: GetState) {
+    const oldMessages = getState().widgets.byId[WIDGET_ID].channelData[channel];
+    const messages = await SLACK_API.fetchRequestMessagesForChannel();
+    if (oldMessages) {
+      // TODO: This fetchRequest needs to be happening in SLACK_API file as
+      // a function named fetchMessagesForChannel--or similar
+      //const messages = await fetchRequest(`${PATH}messages/${channel}`);
+      return;
+    }
+    dispatch({
+      channel,
+      messages,
+      type: 'RECEIVED_MESSAGES_FOR_CHANNEL',
+    });
+  };
+}
 export function processNewMessages(newMessageData: {[string]: ?{[Id]: {[Id]: MessageType}}}) {
   return {
     messages: newMessageData,
@@ -27,34 +65,5 @@ export function selectChannel(channel: string) {
   return {
     channel,
     type: 'SELECT_CHANNEL',
-  };
-}
-
-/* eslint func-names: ["error", "never"] */
-export function fetchChannels() {
-  return async function (dispatch: Dispatch) {
-    // const response = await fetchRequest(`${PATH}channels`);
-    const channels = await SLACK_API.fetchRequestChannels();
-    dispatch({
-      channels,
-      type: 'RECEIVED_CHANNEL_LIST',
-    });
-  };
-}
-
-export function fetchMessagesForChannel(channel: string) {
-  return async function (dispatch: Dispatch, getState: GetState) {
-    const oldMessages = getState().channelData[channel];
-    const messages = await SLACK_API.fetchRequestMessagesForChannel();
-    if (oldMessages) {
-      // Don't fetch again if we already have messages.
-      return;
-    }
-    // Mark that we have messages to avoid fetching multiple times.
-    dispatch({
-      channel,
-      messages,
-      type: 'RECEIVED_MESSAGES_FOR_CHANNEL',
-    });
   };
 }
