+++ src/Constants/index.js
@@ -1,7 +1,9 @@
 // @flow
 
-//export const slackConnectHref = 'https://slack.com/oauth/authorize?scope=incoming-webhook&client_id=218633036018.221954264375';
+const slackConnectHref = process.env.SLACK_CONNECT_HREF;
 
 // export default slackConnectHref;
 
 export const WIDGET_ID = 'slack';
+
+//export const slackConnectHref = 'https://slack.com/oauth/authorize?scope=incoming-webhook&client_id=218633036018.221954264375';
